# 49_Se7enSquare-main/back-end/docs/swagger.json
{
  "openapi": "3.0.0",
  "paths": {
    "/users": {
      "get": {
        "operationId": "UsersController_findAll",
        "summary": "Get all users",
        "parameters": [
          {
            "name": "x-role",
            "in": "header",
            "description": "Role for RBAC: admin | moderator | user",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/UserDto"
                  }
                }
              }
            }
          }
        },
        "tags": [
          "users"
        ]
      },
      "post": {
        "operationId": "UsersController_create",
        "summary": "Create user",
        "parameters": [
          {
            "name": "x-role",
            "in": "header",
            "description": "Role for RBAC: admin | moderator | user",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateUserDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserDto"
                }
              }
            }
          }
        },
        "tags": [
          "users"
        ]
      }
    },
    "/users/{id}": {
      "get": {
        "operationId": "UsersController_findOne",
        "summary": "Get one user by id",
        "parameters": [
          {
            "name": "x-role",
            "in": "header",
            "description": "Role for RBAC: admin | moderator | user",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserDto"
                }
              }
            }
          }
        },
        "tags": [
          "users"
        ]
      },
      "patch": {
        "operationId": "UsersController_update",
        "summary": "Update user",
        "parameters": [
          {
            "name": "x-role",
            "in": "header",
            "description": "Role for RBAC: admin | moderator | user",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateUserDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserDto"
                }
              }
            }
          }
        },
        "tags": [
          "users"
        ]
      },
      "delete": {
        "operationId": "UsersController_remove",
        "summary": "Delete user",
        "parameters": [
          {
            "name": "x-role",
            "in": "header",
            "description": "Role for RBAC: admin | moderator | user",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "example": {
                    "message": "User 4 deleted"
                  }
                }
              }
            }
          }
        },
        "tags": [
          "users"
        ]
      }
    },
...
  },
  "info": {
    "title": "Se7enSquare Backend API",
    "description": "In-memory NestJS backend for frontend integration and academic evaluation",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "x-role": {
        "type": "apiKey",
        "in": "header",
        "name": "x-role",
        "description": "RBAC role header: admin | moderator | user"
      }
    },
    "schemas": {
      "UserDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "number",
            "example": 1
          },
          "username": {
            "type": "string",
            "example": "player01"
          },
          "email": {
            "type": "string",
            "example": "player@example.com"
          },
          "role": {
            "type": "string",
            "enum": [
              "admin",
              "moderator",
              "user"
            ],
            "example": "user"
          },
          "bio": {
            "type": "string",
            "example": "Casual and competitive gamer"
          }
        },
        "required": [
          "id",
          "username",
          "email",
          "role"
        ]
      },
      "CreateUserDto": {
        "type": "object",
        "properties": {
          "username": {
            "type": "string",
            "minLength": 3,
            "maxLength": 30,
            "example": "newplayer"
          },
          "email": {
            "type": "string",
            "example": "newplayer@example.com"
          },
          "role": {
            "type": "string",
            "enum": [
              "admin",
              "moderator",
              "user"
            ],
            "example": "user"
          },
          "bio": {
            "type": "string",
            "minLength": 5,
            "maxLength": 160,
            "example": "Loves MMORPG and FPS."
          }
        },
        "required": [
          "username",
          "email",
          "role"
        ]
      },
      "UpdateUserDto": {
        "type": "object",
        "properties": {
          "username": {
            "type": "string",
            "minLength": 3,
            "maxLength": 30,
            "example": "newplayer"
          },
          "email": {
            "type": "string",
            "example": "newplayer@example.com"
          },
          "role": {
            "type": "string",
            "enum": [
              "admin",
              "moderator",
              "user"
            ],
            "example": "user"
          },
          "bio": {
            "type": "string",
            "minLength": 5,
            "maxLength": 160,
            "example": "Loves MMORPG and FPS."
          }
        }
      },
...
    }
  }
}